Test order of conferences
