+++ scripts/reorderConferencesByDate.js
@@ -1,14 +1,11 @@
 // Reorder a file by running (from the scripts folder)
 const fs = require('fs');
-const sortBy = require('lodash/sortBy');
-const parse = require('date-fns/parse');
 const conferenceReader = require('./utils/conferenceReader');
+const orderConferences = require('./utils/orderConferences');
 
 const BASE_DIR = 'conferences';
 
-const conferencesJSON = conferenceReader();
-
-const propertyOrder = require('../config/validFields');
+const conferencesJSON = conferenceReader(true);
 
 Object.keys(conferencesJSON).forEach(year => {
     Object.keys(conferencesJSON[year]).forEach(topic => {
@@ -18,21 +15,7 @@ Object.keys(conferencesJSON).forEach(year => {
         if (!Array.isArray(conferences)) {
             return;
         }
-        const sortedConfs = sortBy(conferences, [
-            conf => parse(conf.startDate, 'yyyy-MM-dd', new Date()).getTime(),
-            conf => parse(conf.endDate || conf.startDate, 'yyyy-MM-dd', new Date()).getTime(),
-            'name'
-        ]);
-
-        const sortedConfByProperties = sortedConfs.map(conference => {
-            const sortedEntry = {};
-            propertyOrder.forEach(property => {
-                sortedEntry[property] = conference[property];
-            });
-            return sortedEntry;
-        });
-
-        fs.writeFile(fileName, JSON.stringify(sortedConfByProperties, null, 2), () => {
+        fs.writeFile(fileName, orderConferences(conferences), () => {
             console.log(`File ${fileName} was successfully reordered`);
         });
     });
